iptables/ip6tables use optimisation
The Netfilter class used to fork/exec iptables or ip6tables every
time new rules were inserted, appended or deleted. The fork/execs
are now consolidated to one at the end of a hook instead of worst
case scenarios of up to eight fork/execs, significantly improving
execution time.

Also included a fix to a SEGFAULT with the portForwarding option.