RDK-30291 Fixed occasional temp mounts
There was an issue where container was created and not run, it was resulting in temp mount being left because cleaning up was done in post start hook, which wasn't run as container wasn't start.
Fixed by checking if cleaning was done on post stop hook which should be run all the times.