Run plugin hooks in order determined by dependencies
DobbyRdkPluginManager now processes plugins for a given hook in the order of plugin dependency.

If plugin A depends on plugin B, hooks for plugin B will be run before hooks for plugin A.
This is inverted (A before B) for postHalt and postStop so that the plugins on which other plugins depend are shut down later.

The dependencies can be defined in the container spec. Each plugin spec now contains a "dependsOn" array.
If, for example, plugin A depends on plugins B and C, then plugin A's spec can look like this:
"a": {
    "required": false,
    "dependsOn": [ "b", "c" ],
    "data": {}
}

Dependency cycles are detected and make DobbyRdkPluginManager abort before it starts running the plugin hooks.