Always add /sys mount
The /sys mount was conditionally compiled out of the OCI config template
and then added by the Networking plugin if network namespacing was enabled.
Because of that, the /sys mount was appended to the end of the mounts array
and processed after the /sys/fs/cgroup mount. As a result, the /sys/fs/cgroup
mount was failing.

Remove the conditional include of the /sys mount from the template and
the mount logic from the Networking plugin. The mount is now always added
in the proper place of the mount config array, i.e. before the /sys/fs/cgroup
mount.