XIONE-6236 Don't wait for failed forks
We shouldn't do waitpid if fork failed, as it
will result in infinite wait. Also why do we
even use waitpid at all? In forkExecRunC there
is implementation based on vfork which is
blocking by defintion. Which should mean that
if we got to this line we already finished
the child process. I guess it wouldn't harm
(at least as long as we check if it succeed
to fork on the first place).