XIONE-6749 Created high priority watchdog wagger
As on some platforms we expirienced heavy load from unidentified source
Dobby got shut down by the watchdog. It is hard to pinpoint which process
is taking those resources, but it looks like this happens during bootup.
This function if run as separate thread will work around the problem by
creating high priority watchdog wagger for the time period where the
issue exists. During this time there will be 2 concurent wagging procedures,
but this doesn't harm.
We should delete this code when we find out the real offender.