LLAMA-8016 Fixed race with veth devices
Fixed the race when one container was closed and second was
started at the same time. It resulted with situation when
container A still held information about some vethX, while
container B was already assigning it to itself, so when
container A postHaltHook was done it removed all traces
of vethX which made container B loose the network.