ONEM-32702 Delay calling mContainerStoppedCb until very end
The mContainerStoppedCb is called too early. When running
containers on read-only mounted dm-veritfy rootfs, those
rootfs are un-mounted when mContainerStoppedCb is called.
But if this is called too early we get error:
Device or resource busy.

So moving the call to mContainerStoppedCb to the very end
of DobbyManager::onChildExit() so that the extra logic about
"tracking any executed processes inside the container" happens
BEFORE calling mContainerStoppedCb.