Added a check for a env var called "DOBBY_INIT_PAUSE" in DobbyInit, if found and set to "1" then a SIGSTOP is raised on the forked process before exec is called.  This allows for things like gdb to be attached and other debug operations to be performed before the main process starts.
This is only intended as a debugging feature / enabler.

In addition I switched from vfork to fork, as vfork has no
real advantage over fork on linux, and it can mess with SIGSTOP.

Lastly also removed some very old debug code from SkyQ days that
reported the inode of the logging pipe.  It was used to debug
a rare issue we had with logging on SkyQ, but has no value on
RDK.